requirements.txt
Django==4.2
parse-it==2023.10.5.16.31
django-boot-rhazes==1.2.0
pre-commit
